Final migration to new config system
