+++ src/main/java/com/github/u3games/eventengine/EventEngineManager.java
@@ -31,7 +31,6 @@ import com.github.u3games.eventengine.adapter.EventEngineAdapter;
 import com.github.u3games.eventengine.ai.NpcManager;
 import com.github.u3games.eventengine.config.BaseConfigLoader;
 import com.github.u3games.eventengine.datatables.BuffListData;
-import com.github.u3games.eventengine.datatables.ConfigData;
 import com.github.u3games.eventengine.datatables.EventData;
 import com.github.u3games.eventengine.datatables.MessageData;
 import com.github.u3games.eventengine.enums.EventEngineState;
@@ -77,8 +76,6 @@ public class EventEngineManager
 			EventEngineAdapter.class.newInstance();
 			LOGGER.info(EventEngineManager.class.getSimpleName() + ": Adapter loaded.");
 			// Load event configs
-			ConfigData.getInstance();
-			LOGGER.info(EventEngineManager.class.getSimpleName() + ": Configs loaded.");
 			BaseConfigLoader.getInstance();
 			LOGGER.info(EventEngineManager.class.getSimpleName() + ": New Configs loaded.");
 			EventData.getInstance();
