+++ src/main/java/com/github/u3games/eventengine/config/BaseConfigLoader.java
@@ -2,7 +2,10 @@ package com.github.u3games.eventengine.config;
 
 import com.github.u3games.eventengine.config.interfaces.EventConfig;
 import com.github.u3games.eventengine.config.model.MainEventConfig;
-import com.github.u3games.eventengine.events.teamvsteam.TvTEventConfig;
+import com.github.u3games.eventengine.events.types.allvsall.AllVsAllEventConfig;
+import com.github.u3games.eventengine.events.types.capturetheflag.CTFEventConfig;
+import com.github.u3games.eventengine.events.types.survive.SurviveEventConfig;
+import com.github.u3games.eventengine.events.types.teamvsteam.TvTEventConfig;
 import com.google.gson.Gson;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
@@ -14,23 +17,25 @@ public class BaseConfigLoader {
 
     private static final String MAIN_CONFIG_PATH = "./config/EventEngine/EventEngine.conf";
     private static final String TVT_CONFIG_PATH = "./config/EventEngine/TeamVsTeam.conf";
+    private static final String CTF_CONFIG_PATH = "./config/EventEngine/CaptureTheFlag.conf";
+    private static final String ALLVSALL_CONFIG_PATH = "./config/EventEngine/AllVsAll.conf";
+    private static final String SURVIVE_CONFIG_PATH = "./config/EventEngine/Survive.conf";
+
 
     private static BaseConfigLoader sInstance;
 
     private MainEventConfig mMainConfig;
     private TvTEventConfig mTvTConfig;
+    private CTFEventConfig mCtfConfig;
+    private AllVsAllEventConfig mAllVsAllConfig;
+    private SurviveEventConfig mSurviveConfig;
 
     private BaseConfigLoader() {
         mMainConfig = (MainEventConfig) loadConfig(MAIN_CONFIG_PATH, MainEventConfig.class);
         mTvTConfig = (TvTEventConfig) loadConfig(TVT_CONFIG_PATH, TvTEventConfig.class);
-
-        System.out.println("Global message " + mMainConfig.getGlobalMessage());
-        System.out.println("Dualbox max allowed " + mMainConfig.getDualbox().getMaxAllowed());
-
-        System.out.println("Reward kill enabled " + mTvTConfig.isRewardKillEnabled());
-        System.out.println("Team Red " + mTvTConfig.getTeamRed().get(0).toString());
-        System.out.println("Team Blue " + mTvTConfig.getTeamBlue().get(0).toString());
-        System.out.println("Protection time " + mMainConfig.getSpawnProtectionTime());
+        mCtfConfig = (CTFEventConfig) loadConfig(CTF_CONFIG_PATH, CTFEventConfig.class);
+        mAllVsAllConfig = (AllVsAllEventConfig) loadConfig(ALLVSALL_CONFIG_PATH, AllVsAllEventConfig.class);
+        mSurviveConfig = (SurviveEventConfig) loadConfig(SURVIVE_CONFIG_PATH, SurviveEventConfig.class);
     }
 
     private EventConfig loadConfig(String path, Class c) {
@@ -49,6 +54,18 @@ public class BaseConfigLoader {
 
     public TvTEventConfig getTvTConfig() {
         return mTvTConfig;
+    }
+
+    public CTFEventConfig getCtfConfig() {
+        return mCtfConfig;
+    }
+
+    public AllVsAllEventConfig getAllVsAllConfig() {
+        return mAllVsAllConfig;
+    }
+
+    public SurviveEventConfig getSurviveConfig() {
+        return mSurviveConfig;
     }
 
     public static BaseConfigLoader getInstance() {
